All of the routers are converted into controllers for clean code, and the basic login, register and acces token generator routers are working now
