+++ package.json
@@ -10,6 +10,7 @@
   "author": "coder.aliraza",
   "license": "ISC",
   "dependencies": {
+    "bcrypt": "^5.0.1",
     "dotenv": "^16.0.1",
     "express": "^4.18.1",
     "express-list-endpoints": "^6.0.0",
