+++ src/api/controllers/userController.js
@@ -1,0 +1,6 @@
+const UserModel = require("../models/User");
+
+exports.allUsers = async (req, res)=>{
+    const users = await UserModel.findAll()
+    res.json(users)
+}
