+++ src/api/models/User.js
@@ -23,5 +23,6 @@ const User = sequelize.define("User", {
     }
 });
 
+
 module.exports = User;
